requirements.txt
Flask==3.1.2
flask-cors==6.0.2
python-dotenv==1.0.1
PyPDF2==3.0.1
google-genai==0.5.0
